Rewrite application for Ruby 3 and Ruby on Rails 6
This "upgrades" the application from Ruby 1.9.3 to version 3, and Ruby
on Rails 3 to version 6, by rewriting it.

This migrates the application without changing the database model. This
appears today as a pretty bad choice to have User belongs_to Couple,
since it forbids one user to track shared expenses with another user.
Should I had to design it differently, I would create "teams"
referencing two users.

This keeps testing with RSpec, and now relies on FactoryBot to build
model instances from factories instead of inline statements within
examples. RSpec examples have been migrated from should to expectations.
This change doesn't add new request specs, and leaves the default
pending examples as they are, to stand out when executing them.

This change brings in RuboCop to verify the code adheres to good
practices maintained by the Ruby community. It purposedly diverges from
the standards where the developer thinks the rule doesn't feature enough
relevant supporting materials. This change fixes all offenses.

This fixes the display of "No expenses yet", whose code was showed and
not the output.

This commit leaves some issues unresolved. There are CSS issues when
displaying the header navigation bar. JavaScript code to close Active
Model errors pop-up doesn't work. This will be worked on, and it is
preferred to commit the working, upgraded version, as is.