Remove an irrelevant, ever-failing Expense spec
This drops the spec which was testing an Expense being invalid with an
amount lower than 0.01. This may had succeeded in the past with other
databases, however it's always failing now with SQLite3. The rationale
is that the database migration instructs to create an amount field with
a precision of 2 decimals. As a consequence, it cannot store a value
lower than 0.01. SQLite3 silently converts the value to 0.01, which the
spec did not set, and its execution fails. In this case, it's an
irrelevant spec, and shall be removed.