Store expense amount up to 7 total digits
This copes with a difference in how SQLite understands the precision and
scale compared to PostgreSQL. With the former precision of 5 and scale
of 2, SQLite stored Expense#amount having a total of 5 digits, including
a maximum of 2 digits as decimal, while PostgreSQL stored 5 digits for
the integer part, *plus* 2 digits for the decimal part.
After the migration of the production environment from PostgreSQL to
SQLite, checksums didn't match. SQLite rounded large amounts such as
10725.48 to 10725, 3052.54 to 3052.5, while 10000 worked fine. This new
version of the migration and the DB schema ensures numbers are stored
up to 99999.99.