Remove pg gem, ensure sqlite is available in prod
Now that the application has been deleted from Scalingo hosting service,
it runs only locally. The pg gem is no longer needed, and this change
removes it from both the Gemfile and Gemfile.lock. On the other hand,
since the application is using sqlite in production environment, it
ensures it is available outside of development and test groups.
Last, the update of the Gemfile.lock also bumped the version of RuboCop.